requirements.txt
web3==7.6.0
python-dotenv==1.0.0